#1700: extract security/screen ShowText branches to security_text
Move the four screen-* prefix handlers (screen-ids-option:,
screen-statistics:, screen-statistics-all, screen-ids-option-detail:)
and the screen, ike, alg, dynamic-address, address-book switch cases
out of server_show.go into server_show_security_text.go, following the
#1043 pattern: bodies moved verbatim with &buf becoming the passed-in
*strings.Builder. Prefix handlers return the response directly (they
were early-returns; the shared buf is empty on entry). The
screenSYNCookieCounterRows helper moves alongside its sole consumers.

server_show.go drops from 2006 to 1554 LOC. ShowText golden guard
passes (byte-identical) and the full pkg/grpcapi suite is green.
